Feat: 세션 아이디값 참조 후 글 생성 및 삭제 추가
CUD 전부 아이디값 참조 후 글생성,수정,삭제

+++ src/main/java/org/example/newsfeed_project/post/service/PostService.java
@@ -4,13 +4,21 @@ package org.example.newsfeed_project.post.service;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.example.newsfeed_project.entity.User;
+import org.example.newsfeed_project.exception.ValidateException;
 import org.example.newsfeed_project.post.dto.CreatedPostRequestDto;
 import org.example.newsfeed_project.post.dto.CreatedPostResponseDto;
 import org.example.newsfeed_project.post.dto.UpdatedPostRequestDto;
 import org.example.newsfeed_project.post.dto.UpdatedPostResponseDto;
 import org.example.newsfeed_project.entity.Post;
 import org.example.newsfeed_project.post.repository.PostRepository;
+import org.example.newsfeed_project.user.repository.UserRepository;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -18,29 +26,40 @@ import org.springframework.stereotype.Service;
 public class PostService {
 
     private final PostRepository postRepository;
+    private final UserRepository userRepository;
 
     //created
-    public CreatedPostResponseDto createdPost (CreatedPostRequestDto createdRequest) {
-        Post post = new Post(createdRequest.getTitle(), createdRequest.getContents());
+    public CreatedPostResponseDto createdPost (Long userId, CreatedPostRequestDto createdRequest) {
+        User user = userRepository.findUserByUserIdOrElseThrow(userId);
+        Post post = new Post(user, createdRequest.getTitle(), createdRequest.getContents());
         Post savePost = postRepository.save(post);
         return new CreatedPostResponseDto(
+                user.getUserName(),
                 savePost.getTitle(),
                 savePost.getContents(),
-                savePost.getUpdatedAt() );
+                savePost.getUpdatedAt());
     }
     //update
     @Transactional
-    public UpdatedPostResponseDto updatePost(Long post_id, UpdatedPostRequestDto updatePostRequest) {
+    public UpdatedPostResponseDto updatePost(Long userId, Long post_id, UpdatedPostRequestDto updatePostRequest) {
+       User user = userRepository.findUserByUserIdOrElseThrow(userId);
        Post post = postRepository.findById(post_id)
                 .orElseThrow(() -> new IllegalArgumentException("해당 게시물이 존재하지 않습니다."));
+       if(userId != post.getUser().getUserId()) {
+           throw new ValidateException("게시글 작성자가 아닙니다.", HttpStatus.UNAUTHORIZED);
+       }
        post.updatedPost(updatePostRequest.getTitle(), updatePostRequest.getContents());
        return new UpdatedPostResponseDto(post.getTitle(), post.getContents(), post.getUpdatedAt());
     }
-    
+
     @Transactional
-    public void deletePost(Long post_id) {
+    public void deletePost(Long userId, Long post_id) {
+        User user = userRepository.findUserByUserIdOrElseThrow(userId);
         Post post = postRepository.findById(post_id)
                 .orElseThrow(() -> new IllegalArgumentException("해당 게시물이 존재하지 않습니다."));
+        if(userId != post.getUser().getUserId()) {
+            throw new ValidateException("게시글 작성자가 아닙니다", HttpStatus.UNAUTHORIZED);
+        }
         postRepository.deleteById(post_id);
     }
 }
